fix(app): split turn-feed tap handlers under flame 400-line cap
Extract GameMapAreaTurnFeedTaps mixin so game_map_area_turn_feed_labels
stays ≤400 physical lines after overseas-profit feed deep-link (Refs #4226).
